Add files via upload
This is a Scenario Planning tool which is help to choose the best case for the Business. For example: - A customer buy a products and he negotiate on the discount on the products. So the sales manager can easily find the best case which would be more appropriate for the business.